+++ CardPayments/src/main/java/com/paypal/android/cardpayments/model/CardResult.kt
@@ -1,14 +1,11 @@
 package com.paypal.android.cardpayments.model
 
 import android.net.Uri
-import com.paypal.android.corepayments.OrderStatus
 
 /**
  * A result returned by [CardClient] when an order was successfully approved with a [Card].
  */
 data class CardResult(
     val orderID: String,
-    val status: OrderStatus? = null,
-    val paymentSource: PaymentSource? = null,
     val deepLinkUrl: Uri? = null
 )