+++ CardPayments/src/test/java/com/paypal/android/cardpayments/CardClientUnitTest.kt
@@ -7,13 +7,10 @@ import com.braintreepayments.api.BrowserSwitchClient
 import com.braintreepayments.api.BrowserSwitchOptions
 import com.braintreepayments.api.BrowserSwitchResult
 import com.braintreepayments.api.BrowserSwitchStatus
-import com.paypal.android.cardpayments.api.ConfirmPaymentSourceResponse
-import com.paypal.android.cardpayments.api.GetOrderInfoResponse
-import com.paypal.android.cardpayments.api.GetOrderRequest
 import com.paypal.android.cardpayments.api.CheckoutOrdersAPI
+import com.paypal.android.cardpayments.api.ConfirmPaymentSourceResponse
 import com.paypal.android.cardpayments.model.CardResult
 import com.paypal.android.cardpayments.model.PaymentSource
-import com.paypal.android.corepayments.ClientIdRepository
 import com.paypal.android.corepayments.OrderStatus
 import com.paypal.android.corepayments.PayPalSDKError
 import com.paypal.android.corepayments.analytics.AnalyticsService
@@ -34,8 +31,6 @@ import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
-import strikt.api.expectThat
-import strikt.assertions.isEqualTo
 
 @ExperimentalCoroutinesApi
 @RunWith(RobolectricTestRunner::class)
@@ -47,7 +42,6 @@ class CardClientUnitTest {
     private val cardRequest = CardRequest(orderID, card, "return_url")
 
     private val checkoutOrdersAPI = mockk<CheckoutOrdersAPI>(relaxed = true)
-    private val clientIdRepository = mockk<ClientIdRepository>(relaxed = true)
     private val analyticsService = mockk<AnalyticsService>(relaxed = true)
     private val confirmPaymentSourceResponse =
         ConfirmPaymentSourceResponse(orderID, OrderStatus.APPROVED)
@@ -83,30 +77,6 @@ class CardClientUnitTest {
     fun `register lifecycle observer on init`() = runTest {
         createCardClient(testScheduler)
         verify(exactly = 1) { activityLifecycle.addObserver(any<CardLifeCycleObserver>()) }
-    }
-
-    @Test
-    fun `approveOrder() notifies listener if error fetching clientID`() = runTest {
-        val error = PayPalSDKError(123, "fake-description")
-        val errorSlot = slot<PayPalSDKError>()
-
-        coEvery { clientIdRepository.fetchClientId() } throws error
-        every {
-            approveOrderListener.onApproveOrderFailure(capture(errorSlot))
-        } answers { errorSlot.captured }
-
-        val sut = createCardClient(testScheduler)
-
-        sut.approveOrder(mockk(relaxed = true), cardRequest)
-        advanceUntilIdle()
-
-        verify {
-            approveOrderListener.onApproveOrderFailure(any())
-        }
-        expectThat(errorSlot.captured) {
-            get { code }.isEqualTo(123)
-            get { errorDescription }.isEqualTo("Error fetching clientID. Contact developer.paypal.com/support.")
-        }
     }
 
     @Test
@@ -167,25 +137,6 @@ class CardClientUnitTest {
     }
 
     @Test
-    fun `handle browser switch result fetches updated order using browser switch metadata`() =
-        runTest {
-            val sut = createCardClient(testScheduler)
-
-            val browserSwitchResult =
-                createBrowserSwitchResult(BrowserSwitchStatus.SUCCESS, approveOrderMetadata)
-            every { browserSwitchClient.deliverResult(activity) } returns browserSwitchResult
-
-            sut.handleBrowserSwitchResult(activity)
-            advanceUntilIdle()
-
-            val orderRequestSlot = slot<GetOrderRequest>()
-            coVerify(exactly = 1) { checkoutOrdersAPI.getOrderInfo(capture(orderRequestSlot)) }
-
-            val orderRequest = orderRequestSlot.captured
-            assertEquals("sample-order-id", orderRequest.orderId)
-        }
-
-    @Test
     fun `handle browser switch result notifies user of success with updated order info`() =
         runTest {
             val sut = createCardClient(testScheduler)
@@ -193,10 +144,6 @@ class CardClientUnitTest {
             val browserSwitchResult =
                 createBrowserSwitchResult(BrowserSwitchStatus.SUCCESS, approveOrderMetadata)
             every { browserSwitchClient.deliverResult(activity) } returns browserSwitchResult
-
-            val response =
-                GetOrderInfoResponse("sample-order-id", OrderStatus.APPROVED, OrderIntent.CAPTURE)
-            coEvery { checkoutOrdersAPI.getOrderInfo(any()) } returns response
 
             sut.handleBrowserSwitchResult(activity)
             advanceUntilIdle()
@@ -221,34 +168,11 @@ class CardClientUnitTest {
         verify(exactly = 1) { approveOrderListener.onApproveOrderCanceled() }
     }
 
-    @Test
-    fun `handle browser switch result notifies listener of get order failure`() =
-        runTest {
-            val sut = createCardClient(testScheduler)
-
-            val browserSwitchResult =
-                createBrowserSwitchResult(BrowserSwitchStatus.SUCCESS, approveOrderMetadata)
-            every { browserSwitchClient.deliverResult(activity) } returns browserSwitchResult
-
-            val error = PayPalSDKError(0, "mock_error_message")
-            coEvery { checkoutOrdersAPI.getOrderInfo(any()) } throws error
-
-            sut.handleBrowserSwitchResult(activity)
-            advanceUntilIdle()
-
-            val errorSlot = slot<PayPalSDKError>()
-            verify(exactly = 1) { approveOrderListener.onApproveOrderFailure(capture(errorSlot)) }
-
-            val capturedError = errorSlot.captured
-            assertEquals("mock_error_message", capturedError.errorDescription)
-        }
-
     private fun createCardClient(testScheduler: TestCoroutineScheduler): CardClient {
         val dispatcher = StandardTestDispatcher(testScheduler)
         val sut = CardClient(
             activity,
             checkoutOrdersAPI,
-            clientIdRepository,
             analyticsService,
             browserSwitchClient,
             dispatcher