+++ CorePayments/src/main/java/com/paypal/android/corepayments/analytics/AnalyticsEventData.kt
@@ -2,7 +2,6 @@ package com.paypal.android.corepayments.analytics
 
 // Ref: https://blog.klipse.tech/databook/2022/06/22/separate-code-from-data.html
 data class AnalyticsEventData(
-    val clientID: String,
     val environment: String,
     val eventName: String,
     val timestamp: Long,