+++ Demo/src/main/java/com/paypal/android/api/model/ClientId.kt
@@ -1,0 +1,5 @@
+package com.paypal.android.api.model
+
+data class ClientId(
+    val value: String
+)